componente InputField creado para FormElement/EC-54
