+++ frontend/src/shared/components/FormElement.jsx
@@ -1,48 +1,10 @@
 import React, { useState } from "react";
 import { useField } from "formik";
 import { FiEye, FiEyeOff } from "react-icons/fi";
+import InputField from "./InputField";
 
 const FormElement = ({ element }) => {
   const { type, name, label, options = [], ...rest } = element;
-  const [field, meta] = useField(name);
-  const [showPassword, setShowPassword] = useState(false);
-
-  const togglePassword = () => setShowPassword((prev) => !prev);
-  const inputType = type === "password" ? (showPassword ? "text" : "password") : type;
-
-  const renderInput = () => (
-    <div className="input-field">
-      {label && <label htmlFor={name}>{label}</label>}
-
-      <div className="input-wrapper">
-        <input id={name} type={inputType} {...field} {...rest} />
-        {type === "password" && (
-          <span className="password-toggle" onClick={togglePassword}>
-            {showPassword ? <FiEyeOff size={20} /> : <FiEye size={20} />}
-          </span>
-        )}
-      </div>
-
-      {meta.touched && meta.error && <div className="error">{meta.error}</div>}
-    </div>
-  );
-
-  const renderSelect = () => (
-    <div className="input-field">
-      {label && <label htmlFor={name}>{label}</label>}
-      <div className="input-wrapper">
-        <select id={name} {...field} {...rest}>
-          <option value="">-- Seleccionar --</option>
-          {options.map((opt, idx) => (
-            <option key={idx} value={opt.value}>
-              {opt.label}
-            </option>
-          ))}
-        </select>
-      </div>
-      {meta.touched && meta.error && <div className="error">{meta.error}</div>}
-    </div>
-  );
 
   switch (type) {
     case "text":
@@ -50,7 +12,7 @@ const FormElement = ({ element }) => {
     case "email":
     case "number":
     case "tel":
-      return renderInput();
+      return <InputField{... element}/>;
 
     case "select":
       return renderSelect();
